话费奖励分享

+++ Mahjong/Classes/mahjong/gameview/shanghai/normal/MahjongView.cpp
@@ -330,7 +330,7 @@ void MahjongView::checkPlayerIpRepetition(){
 
 int  MahjongView::getNumbersByPoker(string pokers){
     std::vector<std::string> myPokers = StringUtil::split(pokers, ",");
-    int sum = myPokers.size()*4;
+    int sum = (int)myPokers.size()*4;
     Vector<Jong*> allKnowPokers;//玩家已知的牌
     for (auto  hero: playerHero->playerHandJongs) {
         allKnowPokers.pushBack(hero);
